Split tests up over more files
